% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_pub_test.R
\name{theme_pub_test}
\alias{theme_pub_test}
\title{Title}
\usage{
theme_pub_test(base_size = 11, base_family = "sans")
}
\arguments{
\item{base_family}{}
}
\value{
returns a standard theme for express project publications based on theme_bw of ggplot2
}
\description{
Title
}
